Literal::Array: implement `#replace`, and `#insert`. add `#prepend` alias (#142)
* add `#prepend` method via alias (`#unshift`)

* Implement `Literal::Array#insert`

This is very similar to how `#push` works, except it happens at a given
`index`, and the implementation and tests are also very similar.

* Implement `Literal::Array#replace`

`#replace` should work with either an Array or Literal::Array. It should
raise an ArgumentError for any other class of argument, and it should
enforce type checking.